Do not create new httpclient for each download
This is a memory leak and can cause crashes on some devices
